refactor: 패키지 이동 및 클래스 명 수정
패키지 간의 관계를 학습하여 컨트롤러 계층, 서비스 계층 등 계층 별로
패키지를 묶어줌, VendingMachineConsole이 어플리케이션 전체를 대표하도록
ApplicationConsole로 이름 수정
