+++ src/main/java/vendingmachine/domain/coin/Coin.java
@@ -1,4 +1,4 @@
-package vendingmachine.coin;
+package vendingmachine.domain.coin;
 
 import java.util.Arrays;
 import java.util.List;
