+++ src/main/java/vendingmachine/domain/item/Item.java
@@ -1,4 +1,4 @@
-package vendingmachine.item;
+package vendingmachine.domain.item;
 
 import java.util.Objects;
